updateTask

+++ migrations/000004_add_category_field_to_tasks_table.up.sql
@@ -1,3 +1,3 @@
 ALTER TABLE task
-    ADD COLUMN IF NOT EXISTS category_id bigserial references categories;
+    ADD COLUMN IF NOT EXISTS category_id int references categories;
